+++ Sources/StreamChatSwiftUI/ChatChannel/MessageList/MessageView.swift
@@ -167,7 +167,7 @@ public struct MessageTextView<Factory: ViewFactory>: View {
                 )
             }
 
-            StreamTextView(message: message)
+            factory.makeAttachmentTextView(options: .init(mesage: message))
                 .padding(.leading, leadingPadding)
                 .padding(.trailing, trailingPadding)
                 .padding(.top, topPadding)
@@ -244,6 +244,16 @@ struct StreamTextView: View {
                 .foregroundColor(textColor(for: message))
                 .font(fonts.body)
         }
+    }
+}
+
+// Options for the attachment text view.
+public class AttachmentTextViewOptions {
+    // The message to display the text for.
+    public let message: ChatMessage
+    
+    public init(mesage: ChatMessage) {
+        self.message = mesage
     }
 }